teamchess_api/requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.6.2
Automat==22.10.0
cffi==1.16.0
channels==4.1.0
channels-redis==4.2.0
chess==1.10.0
constantly==23.10.4
cryptography==42.0.7
daphne==4.1.2
Django==4.2
django-appconf==1.0.6
django-cryptography==1.1
django-environ==0.10.0
djangorestframework==3.15.1
environ==1.0
hyperlink==21.0.0
idna==3.7
incremental==22.10.0
msgpack==1.0.8
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyOpenSSL==24.1.0
redis==5.0.4
service-identity==24.1.0
six==1.16.0
sqlparse==0.5.0
Twisted==24.3.0
txaio==23.1.1
typing_extensions==4.11.0
zope.interface==6.4
